@@ -10,11 +10,23 @@
   },
   "dependencies": {
     "@vueuse/core": "^13.5.0",
+    "axios": "^1.10.0",
+    "body-parser": "^2.2.0",
     "chart.js": "^3.9.1",
+    "cors": "^2.8.5",
+    "dayjs": "^1.11.13",
+    "dotenv": "^17.2.0",
     "echarts": "^5.6.0",
-    "echarts-gl": "^2.0.9",
+    "electron": "^37.2.4",
     "express": "^5.1.0",
+    "jsonwebtoken": "^9.0.2",
     "lucide-vue-next": "^0.525.0",
+    "multer": "^2.0.2",
+    "mysql2": "^3.14.2",
+    "nedb": "^1.8.0",
+    "node-cron": "^4.2.1",
+    "save-dev": "^0.0.1-security",
+    "ts-node": "^10.9.2",
     "vue": "^3.4.21",
     "vue-chartjs": "^4.1.2",
     "vue-router": "^4.5.1",
@@ -22,10 +34,11 @@
   },
   "devDependencies": {
     "@iconify-json/mdi": "^1.2.3",
+    "@types/express": "^5.0.3",
     "@vitejs/plugin-vue": "^5.0.4",
     "electron": "^30.0.1",
     "electron-builder": "^24.13.3",
-    "typescript": "^5.2.2",
+    "typescript": "^5.8.3",
     "unplugin-icons": "^22.1.0",
     "unplugin-vue-components": "^28.8.0",
     "vite": "^5.1.6",
