@@ -1,19 +1,19 @@
 package com.cowrite.project.controller;
 
 import com.cowrite.project.common.ApiResponse;
-import com.cowrite.project.model.entity.Register;
+import com.cowrite.project.common.context.AuthContext;
+import com.cowrite.project.model.dto.user.UserEmailRequest;
 import com.cowrite.project.model.entity.User;
+import com.cowrite.project.model.vo.UserVO;
 import com.cowrite.project.service.UserService;
-import com.cowrite.project.utils.CodeCreateUtils;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
-import javax.mail.MessagingException;
-import java.util.Collection;
+import javax.servlet.http.HttpServletRequest;
 
 import static com.cowrite.project.common.enums.ResponseCodeEnum.SYSTEM_ERROR;
 
@@ -33,11 +33,9 @@ public class UserController {
     private static final Logger log = LoggerFactory.getLogger(UserController.class);
 
     private final UserService userService;
-    private final CodeCreateUtils codeCreateUtils;
 
-    public UserController(UserService userService, CodeCreateUtils codeCreateUtils) {
+    public UserController(UserService userService) {
         this.userService = userService;
-        this.codeCreateUtils = codeCreateUtils;
     }
 
     /**
@@ -54,59 +52,89 @@ public class UserController {
         }
     }
 
-    @PostMapping("/create")
-    @ApiOperation("创建用户")
-    public ApiResponse<Boolean> createUser(@RequestBody Register register) {
-        Boolean codeVerified = codeCreateUtils.verifyCode(register.getEmail(), register.getCode());
-        User user = new User(register.getName(),
-                register.getEmail(),
-                register.getPassword(),
-                register.getAvatarUrl(),
-               "ACTIVE");
+    /**
+     * 邮箱注册账号
+     */
+    @PostMapping("/register/email")
 
-        if (!codeVerified) {
-            return ApiResponse.error("验证码错误或已过期");
+
+    @ApiOperation("邮箱注册账号")
+    public ApiResponse<UserVO> registerEmail(@RequestBody UserEmailRequest userEmailRequest, HttpServletRequest request) {
+        if (userEmailRequest.getCode().isEmpty() || userEmailRequest.getEmail().isEmpty()){
+            return ApiResponse.error("注册信息不全");
         }
+        return ApiResponse.success(userService.registerByEmail(userEmailRequest, request));
+    }
 
-        Boolean isSuccess = userService.save(user);
-        if (isSuccess) {
-            return ApiResponse.success(true);
+    /**
+     * 根据token获取信息
+     */
+    @GetMapping("/info")
+    @ApiOperation("根据token获取角色信息")
+    public ApiResponse<UserVO> getUserInfoByToken() {
+        User currentUser = userService.getCurrentUser();
+        if (currentUser != null){
+            currentUser.setPassword(null);
+            return ApiResponse.success(UserVO.convertToUserVO(currentUser));
         } else {
-            return ApiResponse.error("创建失败");
+            return ApiResponse.error(SYSTEM_ERROR.code(), "用户不存在");
         }
     }
 
-    @GetMapping("/delete/{id}")
-    @ApiOperation("删除用户")
-    public ApiResponse<Boolean> deleteUser(@PathVariable Long id){
-        Boolean isSuccess  = userService.removeById(id);
-        if(isSuccess) {
-            return ApiResponse.success(true);
-        }else {
-            return ApiResponse.error(SYSTEM_ERROR.code(),"删除失败");
+    /**
+     * 邮箱登录
+     */
+    @PostMapping("/login/email")
+    @ApiOperation("邮箱登录账号")
+    public ApiResponse<UserVO> loginByEmail(@RequestBody UserEmailRequest userEmailRequest, HttpServletRequest request) {
+        return ApiResponse.success(userService.loginByEmail(userEmailRequest, request));
+    }
+
+    /**
+     * 用户退出登录
+     */
+    @PostMapping("/logout")
+    @ApiOperation("用户退出登录")
+    public ApiResponse<String> logout() {
+        String currentToken = AuthContext.getCurrentToken();
+        userService.logoutUser(currentToken);
+        return ApiResponse.success("退出登录成功");
+    }
+
+    /**
+     * 上传头像
+     */
+    @PostMapping("/upload-avatar")
+    @ApiOperation("上传头像")
+    public ApiResponse<String> uploadAvatar(
+            @RequestParam("file") MultipartFile file) {
+        String avatarUrl = userService.uploadAvatar(file);
+        return ApiResponse.success(avatarUrl);
+    }
+
+    /**
+     * 普通用户修改自己的信息
+     */
+    @PutMapping("/update")
+    @ApiOperation("普通用户修改自己的信息")
+    public ApiResponse<String> updateUser(@RequestBody User user) {
+        try {
+            return ApiResponse.success(userService.updateUserInfo(user));
+        } catch (IllegalArgumentException e) {
+            return ApiResponse.error(SYSTEM_ERROR.code(), e.getMessage());
+        } catch (Exception e) {
+            log.error("Unexpected error occurred while updating user info", e);
+            return ApiResponse.error(SYSTEM_ERROR.code(), "服务器内部错误");
         }
     }
 
-    @PostMapping("/update")
-    @ApiOperation("更新用户信息")
-    public ApiResponse<Boolean> updateUser(@RequestBody User user){
-        Boolean isSuccess =userService.updateUser(user);
-        if(isSuccess) {
-            return ApiResponse.success(true);
-        }else {
-            return ApiResponse.error(SYSTEM_ERROR.code(),"更新失败");
-        }
-    }
-
-    @GetMapping("/find_user/{id}")
-    @ApiOperation("查找用户")
-    public ApiResponse<User> findUser(@PathVariable Long id) {
-        User user = userService.getById(id);
-        if (user!= null){
-            return ApiResponse.success(user);
-        }else{
-            return ApiResponse.error(SYSTEM_ERROR.code(),"查找失败");
-        }
-
+    /**
+     * 注销用户
+     */
+    @PostMapping("/delete-account")
+    @ApiOperation("注销用户")
+    public ApiResponse<Void> deleteAccount() {
+        userService.requestAccountDeletion();
+        return ApiResponse.success(null);
     }
 }
