@@ -1,12 +1,12 @@
 package com.cowrite.project.service;
 
 import com.baomidou.mybatisplus.extension.service.IService;
-import com.cowrite.project.model.entity.DocumentTag;
+import com.cowrite.project.model.entity.EntityTag;
 
 /**
  * DocumentTag 服务接口
  * @author Hibiscus-code-generate
  */
-public interface DocumentTagService extends IService<DocumentTag> {
+public interface EntityTagService extends IService<EntityTag> {
 
 }
