@@ -2,8 +2,8 @@ package com.cowrite.project.service.impl;
 
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.cowrite.project.mapper.DocumentTagMapper;
-import com.cowrite.project.model.entity.DocumentTag;
-import com.cowrite.project.service.DocumentTagService;
+import com.cowrite.project.model.entity.EntityTag;
+import com.cowrite.project.service.EntityTagService;
 import org.springframework.stereotype.Service;
 
 /**
@@ -11,6 +11,6 @@ import org.springframework.stereotype.Service;
  * @author Hibiscus-code-generate
  */
 @Service
-public class DocumentTagServiceImpl extends ServiceImpl<DocumentTagMapper, DocumentTag> implements DocumentTagService {
+public class EntityTagServiceImpl extends ServiceImpl<DocumentTagMapper, EntityTag> implements EntityTagService {
 
 }
