@@ -1,5 +1,5 @@
 <template>
-  <div class="container">
+  <div class="outline-editor">
     <draggable
         v-model="localItems"
         item-key="id"
@@ -15,25 +15,25 @@
         <transition name="list" mode="out-in">
           <div
               :key="element.id"
-              class="item"
-              :style="{ paddingLeft: `${element.level * 1.5}rem` }"
+              class="outline-item"
+              :style="{ paddingLeft: `${element.level * 1.75}rem` }"
           >
             <div
                 v-for="i in element.level"
                 :key="i"
-                class="level-line"
-                :style="{ left: `${(i - 1) * 1.5}rem` }"
+                class="level-indicator"
+                :style="{ left: `${(i - 1) * 1.75}rem` }"
             />
             <div
                 class="item-content"
                 @click="() => emit('select', element)"
                 @contextmenu.prevent="openContextMenu($event, element)"
             >
-              <ChevronRight class="icon" />
-              <FileText class="icon icon-blue" />
+              <ChevronRight class="icon arrow" />
+              <FileText class="icon doc-icon" />
               <span class="title">{{ element.title }}</span>
               <button class="add-btn" @click.stop="() => addChild(element)">
-                <Plus class="icon small" />
+                <Plus class="icon plus-icon" />
               </button>
             </div>
           </div>
@@ -41,16 +41,23 @@
       </template>
     </draggable>
 
+    <!-- 上下文菜单 -->
     <div
         v-if="contextMenu.visible"
         class="context-menu"
         :style="{ top: `${contextMenu.y}px`, left: `${contextMenu.x}px` }"
-        @click="contextMenu.visible = false"
+        @click.stop="contextMenu.visible = false"
     >
       <ul>
-        <li class="menu-item" @click="renameItem(contextMenu.item)">重命名</li>
-        <li class="menu-item" @click="copyLink(contextMenu.item)">复制链接</li>
-        <li class="menu-item" @click="deleteItem(contextMenu.item)">删除</li>
+        <li class="menu-item" @click="renameItem(contextMenu.item)">
+          <i class="iconfont icon-edit" /> 重命名
+        </li>
+        <li class="menu-item" @click="copyLink(contextMenu.item)">
+          <i class="iconfont icon-link" /> 复制链接
+        </li>
+        <li class="menu-item danger" @click="deleteItem(contextMenu.item)">
+          <i class="iconfont icon-delete" /> 删除
+        </li>
       </ul>
     </div>
   </div>
@@ -61,7 +68,6 @@ import { ref, watch, defineProps, defineEmits } from 'vue'
 import draggable from 'vuedraggable'
 import { FileText, ChevronRight, Plus } from 'lucide-vue-next'
 
-// 类型定义
 interface DocumentItem {
   id: string
   title: string
@@ -81,18 +87,14 @@ const emit = defineEmits<{
 
 const localItems = ref<DocumentItem[]>([...props.items])
 
-watch(
-    () => props.items,
-    (newVal) => {
-      localItems.value = [...newVal]
-    }
-)
+watch(() => props.items, (newVal) => {
+  localItems.value = [...newVal]
+})
 
 watch(localItems, (val) => emit('update:items', val))
 
-// 拖拽记录
+// 拖拽逻辑
 let dragStartX = 0
-
 const onDragStart = (e: DragEvent) => {
   dragStartX = e.clientX
 }
@@ -105,14 +107,11 @@ const onDragChange = (e: any) => {
   const moved = localItems.value.splice(oldIndex, 1)[0]
   localItems.value.splice(newIndex, 0, moved)
 
-  // 每移动 24px 水平距离 = 一个缩进层级
-  const levelOffset = Math.round(deltaX / 24)
-
+  const levelOffset = Math.round(deltaX / 28) // 调整缩进灵敏度
   const newLevel = Math.min(4, Math.max(0, moved.level + levelOffset))
   moved.level = newLevel
 }
 
-// 控制是否允许移动（默认允许）
 const onMove = () => true
 
 // 添加子节点
@@ -144,7 +143,6 @@ const openContextMenu = (event: MouseEvent, item: DocumentItem) => {
   }
 }
 
-// 菜单功能
 const renameItem = (item: DocumentItem | null) => {
   if (!item) return
   const newTitle = prompt('请输入新标题', item.title)
@@ -155,112 +153,209 @@ const renameItem = (item: DocumentItem | null) => {
 const copyLink = (item: DocumentItem | null) => {
   if (!item) return
   const url = `${window.location.origin}/docs/${item.id}`
-  navigator.clipboard.writeText(url).then(() => alert('已复制链接'))
+  navigator.clipboard.writeText(url).then(() => {
+    console.log('链接已复制')
+  })
   contextMenu.value.visible = false
 }
 
 const deleteItem = (item: DocumentItem | null) => {
   if (!item) return
-  localItems.value = localItems.value.filter((i) => i.id !== item.id)
+  if (confirm(`确定要删除 "${item.title}" 吗？`)) {
+    localItems.value = localItems.value.filter((i) => i.id !== item.id)
+  }
   contextMenu.value.visible = false
 }
 </script>
 
 <style scoped>
-.container {
+.outline-editor {
   position: relative;
+  padding: 0.5rem;
+  background-color: #faf9ff;
+  border-radius: 12px;
+  height: calc(100vh - 4rem); /* 确保高度适应布局 */
+  overflow-y: auto;
+  scrollbar-width: thin;
+  scrollbar-color: #d9cdf9 transparent;
 }
-.item {
+
+.outline-editor::-webkit-scrollbar {
+  width: 6px;
+}
+
+.outline-editor::-webkit-scrollbar-thumb {
+  background-color: #d9cdf9;
+  border-radius: 3px;
+}
+
+.outline-item {
   position: relative;
   user-select: none;
-  transition: all 0.2s;
+  transition: all 0.2s ease;
+  margin-bottom: 2px;
 }
-.level-line {
+
+.level-indicator {
   position: absolute;
   top: 0;
   left: 0;
   height: 100%;
   width: 1px;
-  background-color: #e5e7eb;
+  background-color: #e8e0f5;
+  transition: background-color 0.2s ease;
 }
+
+.outline-item:hover .level-indicator {
+  background-color: #d9cdf9;
+}
+
 .item-content {
   display: flex;
   align-items: center;
-  gap: 4px;
-  margin-bottom: 4px;
-  padding: 6px;
-  border-radius: 4px;
+  gap: 8px;
+  padding: 10px 12px;
+  border-radius: 8px;
   cursor: pointer;
-  transition: background-color 0.2s ease;
+  transition: all 0.2s ease;
+  background-color: white;
+  box-shadow: 0 1px 2px rgba(148, 108, 230, 0.05);
+  border: 1px solid #f0e9ff;
 }
+
 .item-content:hover {
-  background-color: #f1f1f1;
+  background-color: #f8f5ff;
+  transform: translateY(-1px);
+  box-shadow: 0 4px 8px rgba(148, 108, 230, 0.1);
+  border-color: #e0d4f9;
 }
+
+.icon {
+  width: 14px;
+  height: 14px;
+  flex-shrink: 0;
+  transition: all 0.2s ease;
+}
+
+.arrow {
+  color: #c0b1e0;
+  transform: rotate(0deg);
+}
+
+.item-content:hover .arrow {
+  color: #946ce6;
+}
+
+.doc-icon {
+  color: #946ce6;
+}
+
 .title {
+  flex: 1;
   font-size: 14px;
   font-weight: 500;
+  color: #333;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
 }
+
 .add-btn {
-  margin-left: auto;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 24px;
+  height: 24px;
   background: none;
   border: none;
+  border-radius: 6px;
+  cursor: pointer;
+  opacity: 0;
+  transition: all 0.2s ease;
+}
+
+.outline-item:hover .add-btn {
+  opacity: 1;
+}
+
+.plus-icon {
   color: #9ca3af;
-  cursor: pointer;
 }
-.add-btn:hover {
-  color: #3b82f6;
+
+.add-btn:hover .plus-icon {
+  color: #946ce6;
 }
-.icon {
-  width: 12px;
-  height: 12px;
-  color: #d1d5db;
-}
-.icon-blue {
-  color: #3b82f6;
-}
-.icon.small {
-  width: 10px;
-  height: 10px;
-}
-.context-menu {
-  position: absolute;
-  z-index: 50;
-  background-color: white;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.15);
-  font-size: 14px;
-  padding: 4px 0;
-}
-.menu-item {
-  padding: 6px 16px;
-  cursor: pointer;
-}
-.menu-item:hover {
-  background-color: #f5f5f5;
-}
+
+/* 拖拽样式 */
 .drag-ghost {
-  opacity: 0.6;
-  background-color: #f0f9ff;
+  opacity: 0.8;
+  background-color: #f3eeff;
   transform: scale(1.02);
-  transition: all 0.2s ease-in-out;
+  box-shadow: 0 4px 12px rgba(148, 108, 230, 0.15);
+  border: 1px solid #d9cdf9 !important;
 }
+
 .drag-chosen {
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-  border: 1px solid #c3dafe;
-  background-color: #ebf4ff;
+  background-color: #f0e9ff;
+  box-shadow: 0 4px 12px rgba(148, 108, 230, 0.1);
+  border: 1px solid #d9cdf9;
 }
+
+/* 上下文菜单 */
+.context-menu {
+  position: fixed;
+  z-index: 1000;
+  background-color: white;
+  border-radius: 12px;
+  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.1);
+  padding: 8px 0;
+  min-width: 160px;
+  border: 1px solid #f0e9ff;
+}
+
+.menu-item {
+  display: flex;
+  align-items: center;
+  padding: 8px 16px;
+  font-size: 13px;
+  color: #333;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.menu-item i {
+  margin-right: 8px;
+  font-size: 14px;
+  color: #946ce6;
+}
+
+.menu-item:hover {
+  background-color: #f8f5ff;
+  color: #946ce6;
+}
+
+.menu-item.danger:hover {
+  color: #ff4d4f;
+  background-color: #fff1f0;
+}
+
+.menu-item.danger i {
+  color: #ff4d4f;
+}
+
+/* 过渡动画 */
 .list-enter-active,
 .list-leave-active {
-  transition: all 0.3s ease;
+  transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
 }
+
 .list-enter-from {
   opacity: 0;
-  transform: translateY(-6px);
+  transform: translateY(-10px);
 }
+
 .list-leave-to {
   opacity: 0;
-  transform: translateY(6px);
+  transform: translateY(10px);
 }
 </style>
-
